telegram-bot/requirements.txt
python-telegram-bot==22.7
anthropic==0.96.0
python-dotenv==1.2.2
